+++ packages/sdks/snippets/vue/src/router/index.ts
@@ -6,6 +6,7 @@ import LivePreview from '@/views/LivePreview.vue';
 import QuickStart from '@/views/QuickStart.vue';
 import BlogArticle from '@/views/blueprints/BlogArticle.vue';
 import Homepage from '@/views/blueprints/Homepage.vue';
+import NavLinks from '@/views/blueprints/NavLinks.vue';
 import ProductDetails from '@/views/blueprints/ProductDetails.vue';
 import { createRouter, createWebHistory } from 'vue-router';
 
@@ -27,6 +28,7 @@ const router = createRouter({
     { path: '/product/category/:handle', component: ProductDetails },
     { path: '/blogs/:handle', component: BlogArticle },
     { path: '/home', component: Homepage },
+    { path: '/landing-page', component: NavLinks },
     // will match everything and put it under `route.params.pathMatch`
     { path: '/:pathMatch(.*)*', component: QuickStart },
   ],