+++ packages/sdks/output/solid/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Builder.io Solid SDK Changelog (@builder.io/sdk-solid)
 
+## 4.2.0
+
+### Minor Changes
+
+- 3f84ee5: Feature: Added support for POST requests in Content HTTP Requests
+
+## 4.1.3
+
+### Patch Changes
+
+- 9417b4a: fix: parsing of localized values evaluated from data bindings
+
+## 4.1.2
+
+### Patch Changes
+
+- 18fdcb9: fix: removed duplicate menu
+
+## 4.1.1
+
+### Patch Changes
+
+- 59cf58a: Add loading=lazy to RawImg component for better perf
+
 ## 4.1.0
 
 ### Minor Changes